requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.1; python_version >= '3.6'
aiosignal==1.2.0; python_version >= '3.6'
async-timeout==4.0.2; python_version >= '3.6'
attrs==22.1.0; python_version >= '3.5'
certifi==2022.6.15.1; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
filelock==3.8.0; python_version >= '3.7'
frozenlist==1.3.1; python_version >= '3.7'
graphql-core==3.2.1; python_version >= '3.6' and python_version < '4'
idna==3.3; python_version >= '3.5'
multidict==6.0.2; python_version >= '3.7'
numpy==1.23.3; python_version >= '3.8'
pandas==1.4.4; python_version >= '3.8'
psutil==5.9.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==2.4.0; python_version >= '3.6'
pymongo==4.2.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.2.1
pyyaml==6.0
requests==2.28.1; python_version >= '3.7' and python_version < '4'
scipy==1.9.1; python_version < '3.12' and python_version >= '3.8'
sgqlc==16.0; python_version >= '3.6'
shinny-structlog==0.0.4; python_version >= '3.6'
simplejson==3.17.6; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tqsdk2==2.1.7
tqsdk==3.3.0
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
websockets==10.3; python_version >= '3.7'
yarl==1.8.1; python_version >= '3.7'
